Made some implicit conversions explicit

+++ componentLibraries/defaultLibrary/Experimental/Signal2DLookupTable.hpp
@@ -96,7 +96,7 @@ namespace hopsan {
                 else
                 {
                     // Make sure that selected data vector is in range
-                    const int nDataCols = mDataFile.getNumDataCols();
+                    const size_t nDataCols = mDataFile.getNumDataCols();
                     if ( !mDataFile.allRowsHaveSameNumCols() || nDataCols != 3 )
                     {
                         addErrorMessage(HString("Wrong number of data columns: ")+to_hstring(nDataCols)+" != 3");
